# Optimize findTransformECC() by removing redundant computations