@asmorkalov 2024-02-01
The PR reduces test image size, but does not optimize the function itself. The PR does not bring enough add-value to be merged.